perf(svar2): take the SVAR1/seqpro import stack off the from_vcf path
`SparseVar2.from_vcf` opened with `from genoray._svar._regions import
_normalize_samples`, and `_normalize_svar2_regions` imported `_normalize_regions`
from the same module *above* its `if regions is None: return []` early return.
That module imports seqpro (and through it numba and llvmlite) and hirola, and
importing it also runs `genoray._svar`'s package __init__ -- the whole SVAR1
stack. So every conversion paid it, even with no `regions=`.

It showed up as a ~2.2s block before the first Rust log line that was
CONSTANT across a 16x change in cohort width and across two different 400+MB
corpora. Nothing computational is that flat; the tell was the invariance, not
the size.

Two changes, both narrow:

- `_normalize_samples` moves to a new dependency-free `genoray._sample_select`.
  It is pure Python (sets, lists, `Path.read_text`) and needed none of that
  stack. `genoray._svar._regions` re-exports it, so `_svar._core` and the
  existing tests are untouched.
- The `_normalize_regions` import moves below the `regions is None` early
  return. It genuinely needs seqpro (`sp.bed.read` for BED input), so it stays
  where it is -- only the ordering changes, and only the no-regions path
  benefits.

Measured on carter-cn-04, 22 GRCh38-length-skewed contigs, 1.4e9 cells,
defaults, best-of-2:

                       preamble      from_vcf        wall     in-call cores
    S=1,000    before     2.53s         5.57s       6.33s          14.6
               after      0.12s         3.23s       3.76s          24.3   (1.68x)
    S=16,000   before     2.17s         6.84s       7.13s          13.1
               after      0.15s         4.35s       4.83s          18.7   (1.48x)

Stores are byte-identical (`b2c6cbbd663654ca`, `47822bc616a52e1e`), and after a
full conversion none of numba/llvmlite/seqpro/pandera/pandas/scipy is resident.

This is also what makes the preceding commit pay off: deferring the mutcat stack
off `import genoray` measured wall-NEUTRAL on its own, because `from_vcf` then
pulled the same numba/seqpro chain back in through `_svar._regions`. The two are
complementary -- neither alone moves wall.
